requirements.txt
numpy==2.1.2
pillow==10.4.0
scikit-learn==1.5.2
